requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
openai>=1.0.0
anthropic>=0.3.11
google-auth>=2.27.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
numpy>=1.26.4
pandas>=2.1.4
setuptools>=69.0.3
wheel>=0.42.0
pip>=24.0
rich>=14.0.0
markdown-it-py>=3.0.0
pygments>=2.19.1
mdurl>=0.1.2
python-docx>=0.8.11


